requirements.txt
aniso8601==10.0.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Deprecated==1.2.18
Flask==2.3.3
Flask-Cors==4.0.0
Flask-HTTPAuth==4.8.0
Flask-Limiter==3.11.0
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
greenlet==3.1.1
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
limits==4.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==1.24.3
ordered-set==4.1.0
packaging==24.2
pandas==1.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
six==1.17.0
SQLAlchemy==2.0.20
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
